+++ include/mujoco/mjxmacro.h
@@ -397,6 +397,9 @@
     X   ( mjtByte, flex_flatskin,         nflex,         1                    ) \
     X   ( int,     flex_bvhadr,           nflex,         1                    ) \
     X   ( int,     flex_bvhnum,           nflex,         1                    ) \
+    X   ( int,     flexedge_J_rownnz,     nflexedge,     1                    ) \
+    X   ( int,     flexedge_J_rowadr,     nflexedge,     1                    ) \
+    X   ( int,     flexedge_J_colind,     nflexedge,     MJ_M(nv)             ) \
     X   ( float,   flex_rgba,             nflex,         4                    ) \
     X   ( float,   flex_texcoord,         nflextexcoord, 2                    )
 
@@ -730,9 +733,6 @@
     X   ( mjtNum,    cinert,            nbody,       10          ) \
     X   ( mjtNum,    flexvert_xpos,     nflexvert,   3           ) \
     X   ( mjtNum,    flexelem_aabb,     nflexelem,   6           ) \
-    X   ( int,       flexedge_J_rownnz, nflexedge,   1           ) \
-    X   ( int,       flexedge_J_rowadr, nflexedge,   1           ) \
-    X   ( int,       flexedge_J_colind, nflexedge,   MJ_M(nv)    ) \
     X   ( mjtNum,    flexedge_J,        nflexedge,   MJ_M(nv)    ) \
     X   ( mjtNum,    flexedge_length,   nflexedge,   1           ) \
     X   ( mjtNum,    bvh_aabb_dyn,      nbvhdynamic, 6           ) \